FAC-627: refactor - Move `CreateAdviceUseCase` to Correct Package
