+++ flickit-advice-engine/src/test/java/org/flickit/assessment/advice/application/port/in/CreateAdviceUseCaseParamTest.java
@@ -2,6 +2,7 @@ package org.flickit.assessment.advice.application.port.in;
 
 import jakarta.validation.ConstraintViolationException;
 import org.flickit.assessment.advice.application.domain.AttributeLevelTarget;
+import org.flickit.assessment.advice.application.port.in.advice.CreateAdviceUseCase;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
